Read Understanding Society Wave 3

Reads and performs basic cleaning operations on the UKHLS third wave. Missing values as detailed below are all set to NA.

ukhls_read_wave3(
  root = c("C:/"),
  file = "HAR_PR/PR/USoc/Data/SN6614_2022_11_29/tab/ukhls",
  full = TRUE
)

Source

University of Essex, Institute for Social and Economic Research. (2022). Understanding Society: Waves 1-12, 2009-2021 and Harmonised BHPS: Waves 1-18, 1991-2009. [data collection]. 17th Edition. UK Data Service. SN: 6614, DOI: 10.5255/UKDA-SN-6614-18

Arguments

root

Character - the root directory.

file

Character - the file path and name.

full

Logical - TRUE if restricting the sample to full interviews only (excluding proxies)

Value

Returns a data table. Note that:

Missing data ("NA", "", "-1", "-2", "-6", "-7", "-9", "-90", "-90.0", "N/A") is replace with NA, except -8 ("don't know") as this is data.

All variable names are converted to lower case.

Each data point is assigned a weight of 1 as there is no weight variable supplied.

A single sampling cluster is assigned.

The probabilistic sampling unit have the year appended to them.

Details

MISSING VALUES

-1 Don't know. When the respondent does not know the answer to a question.

-2 Refused: When the respondent refuses to answer a question.

-7 Proxy: A question not included in the subset of questions asked of proxy respondents. obtained or not attempted.

-8 Not applicable: Used to signify that a particular variable did not apply to a given respondent usually because of internal routing. For example, men in women only questions.

-9 Missing by error or implausible answer.